+++ SandboxiePlus/MiscHelpers/Common/TreeItemModel.cpp
@@ -7,15 +7,6 @@ QString QTreeViewEx::m_ResetColumns = "Reset Columns";
 
 bool CTreeItemModel::m_DarkMode = false;
 
-#if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
-bool operator < (const QVariant& l, const QVariant& r)
-{
-	auto ret = QVariant::compare(l, r);
-	Q_ASSERT(ret != QPartialOrdering::Unordered);
-	return ret == QPartialOrdering::Less;
-}
-#endif
-
 CTreeItemModel::CTreeItemModel(QObject *parent)
 : QAbstractItemModelEx(parent)
 {