+++ SandboxiePlus/SandMan/main.cpp
@@ -48,7 +48,7 @@ int main(int argc, char *argv[])
 		}
 	}
 	// use a shared setting location when used in a business environment for easier administration
-	theConf = new CSettings(ConfDir, "Xanasoft", "Sandboxie-Plus");
+	theConf = new CSettings(ConfDir, "Sandboxie-Plus", "Xanasoft");
 
 #ifndef _DEBUG
 	InitMiniDumpWriter(QString("SandMan-v%1").arg(CSandMan::GetVersion()).toStdWString().c_str() , QString(theConf->GetConfigDir()).replace("/", "\\").toStdWString().c_str());